feat: implement adding multiple users to a project
